+++ Pod/FlashLightCD.h
@@ -5,8 +5,8 @@ class FlashLightCD : public AnimationCD
 {
 public:
 	// Init
-	FlashLightCD(Animation _animation, int _x, int _y);
-	FlashLightCD(int _x, int _y);
+	FlashLightCD(Animation _animation, float _x, float _y);
+	FlashLightCD(float _x, float _y);
 
 	void Update(float _dt);
 	void Draw(LPDIRECT3DTEXTURE9 _texture);
